Remove submódulo acoes
