--- dominio/alertas/acoes/dao.py
@@ -1,5 +1,0 @@
-from django.conf import settings
-
-from dominio.dao import GenericDAO
-
-
